feat: mypage 폴더위치 변경 및 여행기록 기능 비활성화
